Solucionando bug del desplazamiento mínimo

+++ app/src/main/java/es/uniovi/eii/contacttracker/location/services/LocationForegroundService.kt
@@ -219,7 +219,7 @@ class LocationForegroundService : Service(){
         return LocationTrackRequest(
                 minInterval = config.minInterval,
                 fastestInterval = config.minInterval,
-                smallestDisplacement = config.smallestDisplacement)
+                smallestDisplacement = config.smallestDisplacement.toFloat())
     }
 
     /**
